+++ src/main/java/com/avab/avab/batch/SchedulerConfig.java
@@ -1,9 +1,14 @@
 package com.avab.avab.batch;
 
+import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
+import com.avab.avab.domain.User;
+import com.avab.avab.repository.UserRepository;
+import com.avab.avab.service.UserService;
 import org.springframework.batch.core.Job;
 import org.springframework.batch.core.JobParameters;
 import org.springframework.batch.core.JobParametersBuilder;
@@ -21,6 +26,7 @@ import com.avab.avab.service.FlowService;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.transaction.annotation.Transactional;
 
 @Configuration
 @EnableScheduling
@@ -33,6 +39,7 @@ public class SchedulerConfig {
 
     private final FlowViewCountService flowViewCountService;
     private final FlowService flowService;
+    private final UserService userService;
 
     // 매 30분 마다 수행
     @Scheduled(cron = "0 */30 * * * *")
@@ -65,5 +72,16 @@ public class SchedulerConfig {
                 });
 
         log.info("플로우 조회수 업데이트 완료");
+    }
+
+
+    @Scheduled(cron = "0 0 0 * * *")
+    public void hardDeleteUser() {
+        log.info("user hard delete 작동");
+
+        LocalDate threshold = LocalDate.now().minusDays(30);
+        userService.hardDeleteOldUser(threshold);
+
+        log.info("user hard delete 완료");
     }
 }
